stuff to support fewer duplicate files, saving my disk space

+++ syncWithS3.sh
@@ -5,6 +5,6 @@
 DEST=m.mr-pc.org/ismir16
 SRC=2016_Proceedings_ISMIR_Electronic
 
-mv $SRC.zip $SRC
+[ -f $SRC.zip ] && mv $SRC.zip $SRC
 aws s3 sync $SRC/ s3://$DEST --grants read=uri=http://acs.amazonaws.com/groups/global/AllUsers
-mv $SRC/$SRC.zip .
+[ -f $SRC/$SRC.zip ] && mv $SRC/$SRC.zip .
